Use Containerd directly
Use a Containerd worker with Buildkit and use Nerdctl to start
containers on Containerd.

The downside for now being that root is now required to run Buildkitd
and the Ayup daemon because I couldn't get rootless Nerdctl to work.

The plus side is that we no longer need to call `docker load`. The
image is available in the containerd store after building, where
nerdctl can use it.

Kubernetes also uses Containerd so it can be supported in the future
by setting the containerd namespace. Then images will be instantly
available to k8s after building.