Use Dockerfile if one is present
If we find during analysis that there is a Dockerfile then just use
that.

Previously we assumed that the program was a Python script so to
achieve this the result of the analysis has to be explicitly passed
down to the build and run stages.

The client also receives the result in the analysis step. In the
future this can be made better use of in the UI.

This allows Ayup to build and run itself albeit not in the most
satisfying way