portForward: Prevent half closed connections by explicitly closing
Previously we were allowing the port forward RPC stream to linger in a
zombie state when one side of the forwarded connection had closed.

So we now forcibly close the connection when reading on one side
returns EOF or similar.

Note: I haven't ruled out the read part of a connection could return
EOF while the write is still open and expecting data. However the Go
stdlib doesn't support partial shutdown of a connection.