rpc: Secure connections and login with libp2p
When a libp2p multi-address is given as the listen address or the
remote host address then a secure connection will be created with
libp2p.

Private keys are automatically created if they are not set in the
environment. Clients are authorized via their peer ID (public
key). There is an interactive login endpoint the client can call to
authorize it on a given server. If the server console is not
accessible clients can be authorized via an env variable.

Everything can be configured with environment variables. We use a go
dotenv file to persist environment variables such as the private key
and authorized clients.