analysis/build: Add assistant plugin
The user can specify an assistant directory with some python code in which is
called in a container with the app directory mounted.

This assistant script can generate or modify code in the app directory
before it is further analysed and ran. Files created or modified by
the assistants are uploaded back to the source directory.

Logs produced by the app are also saved and mounted in the assistant container.