Run the app process inside Buildkit
Instead of exporting an image to the Containerd store then running it
with Nerdctl. This runs the application inside a container created by
Buildkitd.

On the plus side this avoids the need for an export step which can
take a long time with large containers. It also removes a couple of
large dependencies.

On the downside buildkit doesn't support port forwarding and a
selection of other features we have to recreate where necessary.