srv: Use detached netns, CNI bridge and forward from the rootless ns
Instead of using the "host's" network inside the rootless kit's
namespace, configure it to use CNI bridge networking so that each
container gets its own IP.

Also use detach netns so that buildkit is mostly running the host's
net namespace. Only containers run in the new net namespace. This
is faster and I found OTEL wasn't working inside the rootless
namespace due to name resolution.

Port forwarding is handled by a process that wraps buildkitd and
attaches to the net ns same as buildkit. We get the most recent
assigned IP address from CNI and hope it's the correct container.