Assistants framework
This commit introduces a number of new features motivated by the need
to make Ayup more flexible and extensible. Primarily it allows modular assistants to perform
tasks such as code generation, build and deployment.

Assistants are containers or internal modules that satisfy the
assistants interface.

They can be chained together into a graph. Each assistant can specify
the next assistant.

To support assistants a '.ayup' state directory has been added. This
contains simple files, such as 'first', which describe some app config
or state. Each app has this directory.

The user can specify that a particular assistant should be used by
running `ay app assistant <name>` which will write the assistant's
name to '.ayup/first'.

Assistant's can be added to an Ayup server using `ay assistants push`,
similar to pushing an app except the assistant is not run immediately.

An assistants souce directory should contain a Dockerfile that at a
minimum specifies the base image. In most cases a cmd/entrypoint will
be required too unless the base image itself is an assistant with a
valid entry point. It should copy any files it needs at runtime into /assistant.

When the assistant is run the following directories are mounted

/in/state
/in/app
/out/state
/out/app

Data from the previous assistant or the initial upload are in /in and
the assistant writes to /out

The */app directories contain the application source. The */state
directories contain files such as log (the previous run's logs) and
next (the next assistant to run). the state directory is in fact the
.ayup directory which is moved out of the app dir during execution.

The app and state dirs are handled differently. The assistant must
copy all files from /in/app to /out/app if it wishes to keep
them. However state files are only modified if they are added to
/out/state.

There are 3 types of assistant: builtin, local and remote. The builtin
assistants are intrinsic to Ayup and can interact directly with
Buildkit and the user. local and remote are very similar, remote means
the assistant was distributed with Ayup and local means the user uploaded
the assistant. In the future remote assistants could mean they are in
a global registry.